refactor and prettify all files Relates #54
