+++ db/connection.js
@@ -1,8 +1,5 @@
 const pg = require("pg");
-const dotenv = require("dotenv");
-
-dotenv.config();
-// console.log(process.env);
+require("dotenv").config();
 
 const db = new pg.Pool({
     connectionString: process.env.DATABASE_URL
