+++ middleware/error.js
@@ -1,9 +1,7 @@
-function handleError(err, req, res, next){
-
+function handleError(err, req, res, next) {
     const errorStatus = err.status || 400;
-    res.status(errorStatus).send({ "error": err });
+    res.status(errorStatus).send({error: err});
     // console.error(err);
-
 }
 
 module.exports = handleError;
