+++ server.js
@@ -1,8 +1,8 @@
 const express = require("express");
 const handleError = require("./middleware/error");
 const auth = require("./middleware/auth");
-const examples = require("./handlers/examples");
-const users = require("./handlers/users");
+const examples = require("./handlers/examples-handlers");
+const users = require("./handlers/users-handlers");
 require("dotenv").config();
 
 const PORT = process.env.PORT || 3000;
@@ -10,9 +10,9 @@ const server = express();
 server.use(express.json());
 
 server.get("/", examples.getAllExamples);
-server.post("/examples", auth, examples.post);
+server.post("/examples", auth, examples.postExample);
 server.get("/examples/:id", examples.getExample);
-server.delete("/examples/:id", auth, examples.del);
+server.delete("/examples/:id", auth, examples.deleteExample);
 server.put("/examples/:id", auth, examples.updateExample); //NEED TO TEST
 
 server.post("/signup", users.signup);
@@ -21,9 +21,7 @@ server.post("/login", users.login);
 server.use(handleError);
 
 if (process.env.PGDATABASE !== "localtest") {
-  server.listen(PORT, () =>
-    console.log(`Listening on http://localhost:${PORT}`)
-  );
+    server.listen(PORT, () => console.log(`Listening on http://localhost:${PORT}`));
 }
 
 module.exports = server;
